requirements.txt
numpy
pandas
opencv-python
tqdm
dataclasses
matplotlib
python-dotenv
uvicorn
torchvision
imutils
scipy
scikit-image
opencv_python
opencv-python-headless
filterpy
scikit-learn
seaborn
plotly
super-gradients
boxmot
pyyaml
pillow
